decouple process monitoring and resource monitoring from self heal ( OFW-2074 )
Implemented a new syscfg variable named process_monitor_interval
which is used to run the process_monitor script (task_health_monitor.sh)
based on the set interval. Also decoupled the process monitor from
resource_monitor.sh script.

Signed off by: Arundhati Ghose <aghose.contractor@libertyglobal.com>

+++ source/dmltad/cosa_selfheal_dml.c
@@ -193,6 +193,7 @@ BOOL SelfHeal_SetParamBoolValue
             system("/usr/ccsp/tad/self_heal_connectivity_test.sh &");
             system("/usr/ccsp/tad/resource_monitor.sh &");
             system("/usr/ccsp/tad/selfheal_aggressive.sh &");
+            system("/usr/ccsp/tad/task_health_monitor.sh &");
         }
         else
         {
@@ -230,6 +231,19 @@ BOOL SelfHeal_SetParamBoolValue
                 CcspTraceWarning(("%s: Aggressive self heal script is not running\n", __FUNCTION__));
             } else {
                 CcspTraceWarning(("%s: Aggressive self heal script\n", __FUNCTION__));
+                rc = sprintf_s(cmd, sizeof(cmd), "kill -9 %s", buf);
+                if(rc < EOK)
+                {
+                    ERR_CHK(rc);
+                }
+                system(cmd);
+            }
+
+            copy_command_output("busybox pidof task_health_monitor.sh", buf, sizeof(buf));
+            if (strlen(buf) == 0) {
+                CcspTraceWarning(("%s: Process Monitor script is not running\n", __FUNCTION__));
+            } else {
+                CcspTraceWarning(("%s: Process Monitor script\n", __FUNCTION__));
                 rc = sprintf_s(cmd, sizeof(cmd), "kill -9 %s", buf);
                 if(rc < EOK)
                 {
@@ -1595,6 +1609,12 @@ ResourceMonitor_GetParamUlongValue
         *puLong = pRescMonitor->AvgMemThreshold;
         return TRUE;
     }
+
+    if (strcmp(ParamName, "X_RDKCENTRAL-COM_ProcessMonitorInterval") == 0)
+    {
+        *puLong = pRescMonitor->ProcessMonIntervalTime;
+        return TRUE;
+    }
     return FALSE;
 }
 
@@ -1705,6 +1725,21 @@ ResourceMonitor_SetParamUlongValue
 	return TRUE;
     }
 
+    if (strcmp(ParamName, "X_RDKCENTRAL-COM_ProcessMonitorInterval") == 0)
+    {
+        if ( pRescMonitor->ProcessMonIntervalTime == uValue )
+        {
+            return TRUE;
+        }
+
+        if (syscfg_set_u_commit(NULL, "process_monitor_interval", uValue) != 0)
+        {
+            CcspTraceWarning(("%s: syscfg_set failed for %s\n", __FUNCTION__, ParamName));
+            return FALSE;
+        }
+        pRescMonitor->ProcessMonIntervalTime = uValue;
+        return TRUE;
+    }
     return FALSE;
 }
